+++ tests/test_validation.py
@@ -21,7 +21,7 @@ import pytest
 import pandas as pd
 from pandas.util.testing import assert_series_equal, assert_frame_equal
 from pdvalidate.validation import ei, \
-                                  pdvalidate as pdv, \
+                                  validate as pdv, \
                                   ValidationWarning